+++ database/staking_pool.go
@@ -19,7 +19,7 @@ ON CONFLICT (one_row_id) DO UPDATE
         height = excluded.height
 WHERE staking_pool.height <= excluded.height`
 
-	_, err := db.Sql.Exec(stmt,
+	_, err := db.SQL.Exec(stmt,
 		pool.BondedTokens.String(),
 		pool.NotBondedTokens.String(),
 		pool.UnbondingTokens.String(),