+++ modules/distribution/module.go
@@ -5,7 +5,7 @@ import (
 
 	distrsource "github.com/forbole/bdjuno/v3/modules/distribution/source"
 
-	"github.com/forbole/juno/v3/modules"
+	"github.com/forbole/juno/v4/modules"
 
 	"github.com/forbole/bdjuno/v3/database"
 )